Resolved issue with history crash.
Problem was caused by the mixture of itemId based history and file path
logic.  Left a TODO to remove the itemId as part of a future change.
Item ID based logic should be at a higher level and not be part of the
kdb-repo which should be more general in nature.