Markdown image regular expression modified and user interface for k-markdown and kd-markdown more closely aligned with that of markdown-editor
Both of these tasks were worked on the same branch, with a reason for
that being that they are reasonably related. The work for each of these
tasks resulted in modification to k-markdown.component.html and
kd-markdown.component.html. Thus, work for both of these tasks has been
included in the same commit.